feat(contextmenu): adding support for shortcut label in menu items
An option called "shortcut" was added to the MenuItem definition. When present, the context menu
will be rendered with the addition of an extra string at the end of the row, which contains the
content of the "shortcut" property. This is useful to help the user identify keyboard shortcuts more
easily.